+++ web/src/containers/LoginPageRoutes.js
@@ -98,6 +98,7 @@ const LoginPageRoutes = ({
             cancelAction={handleLogout}
             errorMessage={errorMessage}
             fetching={fetching}
+            secondaryButtonText="Back to Login"
           />
         </Route>